1 [VER] 1.22.8 2 [CHG] change some parameter type uint to long

+++ main/product/Codes/DevLib.Timers/NativeAPI/NativeMethodsHelper.cs
@@ -17,9 +17,9 @@ namespace DevLib.Timers.NativeAPI
         /// Method GetLastInputTime.
         /// </summary>
         /// <returns>The time in milliseconds since last user input.</returns>
-        public static uint GetLastInputTime()
+        public static long GetLastInputTime()
         {
-            uint idleTime = 0;
+            long idleTime = 0;
 
             LASTINPUTINFO lastInputInfo = new LASTINPUTINFO();
             lastInputInfo.cbSize = (uint)Marshal.SizeOf(lastInputInfo);
